+++ samletkodexm1/samletkodexm1/WindowControl.h
@@ -6,12 +6,12 @@
 #include "Aktuator.h"
 
 // Thresholds for temperature, humidity, and CO2
-extern double CO2_OPEN_THRESHOLD; 
-extern double CO2_CLOSE_THRESHOLD; 
-extern double TEMP_OPEN_THRESHOLD;
-extern double TEMP_CLOSE_THRESHOLD; 
-extern double HUMIDITY_OPEN_THRESHOLD; 
-extern double HUMIDITY_CLOSE_THRESHOLD;
+extern float CO2_OPEN_THRESHOLD; 
+extern float CO2_CLOSE_THRESHOLD; 
+extern float TEMP_OPEN_THRESHOLD;
+extern float TEMP_CLOSE_THRESHOLD; 
+extern float HUMIDITY_OPEN_THRESHOLD; 
+extern float HUMIDITY_CLOSE_THRESHOLD;
 // Extern declaration for the global window state
 extern bool isWindowOpen;
 // Function declarations
